refactor : 정산 관련 코드 domain 디렉토리 하위로 구조 변경
