+++ src/main/java/space/space_spring/dao/CommentLikeDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import space.space_spring.entity.Comment;
 import space.space_spring.entity.CommentLike;
-import space.space_spring.entity.User;
+import space.space_spring.domain.user.model.entity.User;
 
 import java.util.Optional;
