+++ src/main/java/space/space_spring/domain/authorization/auth/service/AuthService.java
@@ -9,8 +9,8 @@ import space.space_spring.domain.authorization.jwt.model.TokenPairDTO;
 import space.space_spring.domain.authorization.jwt.model.TokenType;
 import space.space_spring.domain.authorization.jwt.repository.JwtRepository;
 import space.space_spring.domain.user.model.PostLoginDto;
-import space.space_spring.entity.RefreshTokenStorage;
-import space.space_spring.entity.User;
+import space.space_spring.domain.authorization.jwt.model.entity.RefreshTokenStorage;
+import space.space_spring.domain.user.model.entity.User;
 import space.space_spring.exception.CustomException;
 import space.space_spring.domain.authorization.jwt.model.JwtLoginProvider;
 import space.space_spring.util.user.UserUtils;
