+++ src/main/java/space/space_spring/domain/authorization/jwt/service/JwtService.java
@@ -8,8 +8,8 @@ import org.springframework.transaction.annotation.Transactional;
 import space.space_spring.domain.authorization.jwt.model.*;
 import space.space_spring.domain.authorization.jwt.repository.JwtRepository;
 import space.space_spring.domain.user.repository.UserRepository;
-import space.space_spring.entity.RefreshTokenStorage;
-import space.space_spring.entity.User;
+import space.space_spring.domain.authorization.jwt.model.entity.RefreshTokenStorage;
+import space.space_spring.domain.user.model.entity.User;
 import space.space_spring.exception.CustomException;
 import space.space_spring.exception.jwt.unauthorized.JwtExpiredTokenException;
 import space.space_spring.exception.jwt.unauthorized.JwtUnauthorizedTokenException;
