+++ src/main/java/space/space_spring/domain/pay/controller/PayController.java
@@ -1,4 +1,4 @@
-package space.space_spring.controller;
+package space.space_spring.domain.pay.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -7,17 +7,17 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import space.space_spring.argumentResolver.jwtLogin.JwtLoginAuth;
 import space.space_spring.argumentResolver.userSpace.CheckUserSpace;
-import space.space_spring.dto.pay.dto.PayReceiveInfoDto;
-import space.space_spring.dto.pay.dto.PayRequestInfoDto;
-import space.space_spring.dto.pay.dto.TotalPayInfoDto;
-import space.space_spring.dto.pay.request.PostPayCompleteRequest;
-import space.space_spring.dto.pay.request.PostPayCreateRequest;
-import space.space_spring.dto.pay.response.*;
-import space.space_spring.entity.PayRequestTarget;
-import space.space_spring.entity.User;
+import space.space_spring.domain.pay.model.dto.PayReceiveInfoDto;
+import space.space_spring.domain.pay.model.dto.PayRequestInfoDto;
+import space.space_spring.domain.pay.model.dto.TotalPayInfoDto;
+import space.space_spring.domain.pay.model.request.PostPayCompleteRequest;
+import space.space_spring.domain.pay.model.request.PostPayCreateRequest;
+import space.space_spring.domain.pay.model.response.*;
+import space.space_spring.domain.pay.model.entity.PayRequestTarget;
+import space.space_spring.domain.user.model.entity.User;
 import space.space_spring.exception.CustomException;
 import space.space_spring.response.BaseResponse;
-import space.space_spring.service.PayService;
+import space.space_spring.domain.pay.service.PayService;
 import space.space_spring.util.pay.PayUtils;
 import space.space_spring.util.user.UserUtils;
 import space.space_spring.util.userSpace.UserSpaceUtils;
