+++ src/main/java/space/space_spring/domain/pay/model/entity/PayRequest.java
@@ -1,7 +1,10 @@
-package space.space_spring.entity;
+package space.space_spring.domain.pay.model.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
+import space.space_spring.domain.user.model.entity.User;
+import space.space_spring.entity.BaseEntity;
+import space.space_spring.entity.Space;
 
 @Getter
 @Entity
