+++ src/main/java/space/space_spring/domain/pay/model/request/PostPayCompleteRequest.java
@@ -1,4 +1,4 @@
-package space.space_spring.dto.pay.request;
+package space.space_spring.domain.pay.model.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
