+++ src/main/java/space/space_spring/domain/pay/model/response/GetReceivePayViewResponse.java
@@ -1,8 +1,8 @@
-package space.space_spring.dto.pay.response;
+package space.space_spring.domain.pay.model.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import space.space_spring.dto.pay.dto.PayReceiveInfoDto;
+import space.space_spring.domain.pay.model.dto.PayReceiveInfoDto;
 
 import java.util.ArrayList;
 import java.util.List;
