+++ src/main/java/space/space_spring/domain/user/repository/UserRepository.java
@@ -3,7 +3,7 @@ package space.space_spring.domain.user.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import space.space_spring.entity.User;
+import space.space_spring.domain.user.model.entity.User;
 
 import java.util.Optional;
