+++ src/main/java/space/space_spring/dto/post/request/CreatePostRequest.java
@@ -5,7 +5,7 @@ import org.springframework.web.multipart.MultipartFile;
 import space.space_spring.entity.Post;
 import space.space_spring.entity.PostImage;
 import space.space_spring.entity.Space;
-import space.space_spring.entity.User;
+import space.space_spring.domain.user.model.entity.User;
 
 import java.util.List;
