+++ src/main/java/space/space_spring/entity/Post.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import space.space_spring.domain.user.model.entity.User;
 
 import java.util.List;
