+++ src/main/java/space/space_spring/util/pay/PayUtils.java
@@ -3,9 +3,9 @@ package space.space_spring.util.pay;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import space.space_spring.dao.PayDao;
-import space.space_spring.entity.PayRequest;
-import space.space_spring.entity.PayRequestTarget;
+import space.space_spring.domain.pay.repository.PayDao;
+import space.space_spring.domain.pay.model.entity.PayRequest;
+import space.space_spring.domain.pay.model.entity.PayRequestTarget;
 import space.space_spring.exception.CustomException;
 
 import static space.space_spring.response.status.BaseExceptionResponseStatus.PAY_REQUEST_NOT_FOUND;
