+++ src/main/java/space/space_spring/util/userSpace/UserSpaceUtils.java
@@ -1,6 +1,5 @@
 package space.space_spring.util.userSpace;
 
-import space.space_spring.entity.User;
 import space.space_spring.entity.UserSpace;
 
 import java.util.Optional;
