+++ src/test/java/space/space_spring/domain/authorization/jwt/service/JwtServiceTest.java
@@ -18,8 +18,8 @@ import space.space_spring.domain.authorization.jwt.model.TokenPairDTO;
 import space.space_spring.domain.authorization.jwt.model.JwtLoginTokenResolver;
 import space.space_spring.domain.authorization.jwt.repository.JwtRepository;
 import space.space_spring.domain.user.repository.UserRepository;
-import space.space_spring.entity.RefreshTokenStorage;
-import space.space_spring.entity.User;
+import space.space_spring.domain.authorization.jwt.model.entity.RefreshTokenStorage;
+import space.space_spring.domain.user.model.entity.User;
 import space.space_spring.entity.enumStatus.UserSignupType;
 import space.space_spring.exception.jwt.unauthorized.JwtExpiredTokenException;
 import space.space_spring.exception.jwt.unauthorized.JwtUnauthorizedTokenException;
