+++ src/test/java/space/space_spring/repository/UserChatRoomDaoTest.java
@@ -9,7 +9,7 @@
 //import space.space_spring.dao.chat.UserChatRoomDao;
 //import space.space_spring.entity.ChatRoom;
 //import space.space_spring.entity.Space;
-//import space.space_spring.entity.User;
+//import space.space_spring.domain.user.model.entity.User;
 //import space.space_spring.entity.UserChatRoom;
 //import space.space_spring.entity.enumStatus.UserSignupType;
 //
