+++ src/test/java/space/space_spring/service/PayServiceTest.java
@@ -7,14 +7,15 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import space.space_spring.dao.PayDao;
-import space.space_spring.dto.pay.dto.PayReceiveInfoDto;
-import space.space_spring.dto.pay.dto.PayRequestInfoDto;
-import space.space_spring.dto.pay.request.PostPayCreateRequest;
-import space.space_spring.entity.PayRequest;
-import space.space_spring.entity.PayRequestTarget;
+import space.space_spring.domain.pay.repository.PayDao;
+import space.space_spring.domain.pay.model.dto.PayReceiveInfoDto;
+import space.space_spring.domain.pay.model.dto.PayRequestInfoDto;
+import space.space_spring.domain.pay.model.request.PostPayCreateRequest;
+import space.space_spring.domain.pay.model.entity.PayRequest;
+import space.space_spring.domain.pay.model.entity.PayRequestTarget;
+import space.space_spring.domain.pay.service.PayService;
 import space.space_spring.entity.Space;
-import space.space_spring.entity.User;
+import space.space_spring.domain.user.model.entity.User;
 import space.space_spring.entity.enumStatus.UserSignupType;
 import space.space_spring.util.space.SpaceUtils;
 import space.space_spring.util.user.UserUtils;
